--- app/src/test/java/com/monkeyteam/chimpagne/smallClassesTest.kt
@@ -1,3 +1,0 @@
-package com.monkeyteam.chimpagne
-
-class smallClassesTest {}
